解决jobevent回传的id，由于int类型长度小于限制的bug已经统一下单新增了服务商的特有参数

+++ weixin-mp/src/main/java/com/riversoft/weixin/mp/event/template/JobFinishedEvent.java
@@ -11,17 +11,17 @@ import com.riversoft.weixin.common.event.EventRequest;
 public class JobFinishedEvent extends EventRequest {
 
     @JsonProperty("MsgID")
-    private int msgId;
+    private Long msgId;
 
     @JsonProperty("Status")
     @JacksonXmlCData
     private String status;
 
-    public int getMsgId() {
+    public Long getMsgId() {
         return msgId;
     }
 
-    public void setMsgId(int msgId) {
+    public void setMsgId(Long msgId) {
         this.msgId = msgId;
     }
